finish

+++ src/js/user-service/userServ.js
@@ -1,6 +1,7 @@
 import { async } from '@firebase/util';
 import { getAuth } from 'firebase/auth';
 import { app } from '../firebase/initFirebase';
+import { notifyFailureMessage } from '../helpers/notifyMessages';
 import { currentFilmId } from '../modals/filmDetailsModal';
 import {
   currentFilmKey,
@@ -9,55 +10,89 @@ import {
   getFilmDataById,
   deleteDataWithDb,
 } from './dataDatebaseFunc';
+// const userId = getCurrentUser().uid;
+// if (!userId) {
+//   return notifyFailureMessage(
+//     'Something went wrong! Maybe you are not Sign In'
+//   );
+// } else {
 
+// }
+const auth = getAuth(app);
+// const userId = getCurrentUser().uid;
 function addFilmToQueque(e) {
   e.preventDefault();
   const addToQueue = document.querySelector('.film-modal__btn--queue');
   const userId = getCurrentUser().uid;
-  const default_label = 'add to queue';
-  console.log(addToQueue.textContent);
-  if (addToQueue.textContent === default_label) {
-    const filmKey = postFilmToDatabase({
-      id: userId,
-      currentFilmId,
-      reference: 'userQueue',
-    });
-    console.log('before getFilmDatat func');
-    getFilmDataById(userId, 'userQueue', filmKey);
-    addToQueue.disabled = true;
-    setTimeout(() => (addToQueue.disabled = false), 2000);
-    addToQueue.textContent = 'Remove from Queue';
+
+  if (!userId) {
+    return notifyFailureMessage(
+      'Something went wrong! Maybe you are not Sign In'
+    );
   } else {
-    deleteFilmFromQueue();
-    addToQueue.textContent = default_label;
-    addToQueue.disabled = true;
-    setTimeout(() => (addToQueue.disabled = false), 2000);
+    const default_label = 'add to queue';
+    console.log(addToQueue.textContent);
+    if (addToQueue.textContent === default_label) {
+      const filmKey = postFilmToDatabase({
+        id: userId,
+        currentFilmId,
+        reference: 'userQueue',
+      });
+      console.log('before getFilmDatat func');
+
+      console.log(getFilmDataById(userId, 'userQueue', filmKey));
+      //
+      //
+
+      //
+      //
+      addToQueue.disabled = true;
+      setTimeout(() => (addToQueue.disabled = false), 2000);
+      addToQueue.textContent = 'Remove from Queue';
+      addToQueue.classList.add('film-modal__btn-queue-active');
+    } else {
+      deleteFilmFromQueue();
+      addToQueue.textContent = default_label;
+      addToQueue.classList.remove('film-modal__btn-queue-active');
+      addToQueue.disabled = true;
+      setTimeout(() => (addToQueue.disabled = false), 2000);
+    }
   }
 }
 function addFilmToWatched(e) {
   e.preventDefault();
   const userId = getCurrentUser().uid;
-
-  const addToWatch = document.querySelector('.film-modal__btn--watched');
-  const default_label = 'add to queue';
-  console.log(addToWatch.textContent);
-  if (addToWatch.textContent === default_label) {
-    const filmKey = postFilmToDatabase({
-      id: userId,
-      currentFilmId,
-      reference: 'userQueue',
-    });
-    console.log('before getFilmDatat func');
-    getFilmDataById(userId, 'userQueue', filmKey);
-    addToWatch.disabled = true;
-    setTimeout(() => (addToWatch.disabled = false), 2000);
-    addToWatch.textContent = 'Remove from Queue';
+  // if (!userId) {
+  //   return notifyFailureMessage(
+  //     'Something went wrong! Maybe you are not Sign In'
+  //   );
+  if (!userId) {
+    return notifyFailureMessage(
+      'Something went wrong! Maybe you are not Sign In'
+    );
   } else {
-    console.log('before deleteFilm func');
-    deleteFilmFromWatched()();
-    addToWatch.textContent = default_label;
-    addToWatch.disabled = true;
-    setTimeout(() => (addToWatch.disabled = false), 2000);
+    const addToWatch = document.querySelector('.film-modal__btn--watched');
+    const default_label = 'add to Watched';
+    console.log(addToWatch.textContent);
+    if (addToWatch.textContent === default_label) {
+      const filmKey = postFilmToDatabase({
+        id: userId,
+        currentFilmId,
+        reference: 'userWatched',
+      });
+
+      getFilmDataById(userId, 'userWatched', filmKey);
+      addToWatch.disabled = true;
+      setTimeout(() => (addToWatch.disabled = false), 2000);
+      addToWatch.textContent = 'Remove from Watched';
+      addToWatch.classList.add('film-modal__btn-watched-active');
+    } else {
+      deleteFilmFromWatched();
+      addToWatch.textContent = default_label;
+      addToWatch.classList.remove('film-modal__btn-watched-active');
+      addToWatch.disabled = true;
+      setTimeout(() => (addToWatch.disabled = false), 2000);
+    }
   }
 }
 /**
@@ -67,7 +102,13 @@ function addFilmToWatched(e) {
  */
 function deleteFilmFromWatched() {
   const userId = getCurrentUser().uid;
-  deleteDataWithDb({ id: userId, currentFilmId, reference: 'userWatched' });
+  if (!userId) {
+    return notifyFailureMessage(
+      'Something went wrong! Maybe you are not Sign In'
+    );
+  } else {
+    deleteDataWithDb({ id: userId, currentFilmId, reference: 'userWatched' });
+  }
 }
 /**
  *
@@ -76,10 +117,15 @@ function deleteFilmFromWatched() {
  */
 function deleteFilmFromQueue() {
   const userId = getCurrentUser().uid;
-  deleteDataWithDb({ id: userId, currentFilmId, reference: 'userQueue' });
+  if (!userId) {
+    return notifyFailureMessage(
+      'Something went wrong! Maybe you are not Sign In'
+    );
+  } else {
+    deleteDataWithDb({ id: userId, currentFilmId, reference: 'userQueue' });
+  }
 }
 function getCurrentUser() {
-  const auth = getAuth(app);
   const user = auth.currentUser;
   if (user !== null) {
     console.log('get user');
@@ -87,6 +133,7 @@ function getCurrentUser() {
     return user;
   }
   console.log('no user');
+  return false;
 }
 // getCurrentUser();
 // console.log(getCurrentUser());
